+++ CHANGELOG.md
@@ -44,6 +44,7 @@ About changelog [here](https://keepachangelog.com/en/1.0.0/)
 - Parse severity predictions keys/values: CADD, Spidex and Revel scores (#53)
 - Parse conservation: GERP, PhastCons and PhyloP (#54)
 - Parse callers (#55)
+- Parse RankResult INFO field (#56)
 ### Changed
 - Structure of parsed entities, to reflect documents saved in the database (#20)
 - Improved and simplified code (#22)